Use winsock2 so that mingw-gcc can get the correct values for multicast.

+++ generic/udp_tcl.c
@@ -7,7 +7,7 @@
  * Written by Xiaotao Wu
  * Last modified: 11/03/2000
  *
- * $Id: udp_tcl.c,v 1.38 2007/04/10 12:25:57 patthoyts Exp $
+ * $Id: udp_tcl.c,v 1.39 2007/04/10 23:36:14 patthoyts Exp $
  ******************************************************************************/
 
 #if defined(_DEBUG) && !defined(DEBUG)
@@ -730,7 +730,7 @@ SocketThread(LPVOID arg)
         
         /* set each socket for select */
         for (statePtr = sockList; statePtr != NULL; statePtr=statePtr->next) {
-            FD_SET(statePtr->sock, &readfds);
+            FD_SET((unsigned int)statePtr->sock, &readfds);
             UDPTRACE("SET sock %d\n", statePtr->sock);
         }
         
@@ -1203,7 +1203,7 @@ UdpMulticast(ClientData instanceData, Tcl_Interp *interp,
     mreq.imr_interface.s_addr = INADDR_ANY;
     if (setsockopt(statePtr->sock, IPPROTO_IP, action,
                    (const char*)&mreq, sizeof(mreq)) < 0) {
-       Tcl_SetObjResult(interp, ErrorToObj("error changing multicast group"));
+        Tcl_SetObjResult(interp, ErrorToObj("error changing multicast group"));
         return TCL_ERROR;
     }
